Multicommands now journaled and replayed on startup.
git-svn-id: https://wush.net/svn/mindalliance@1321 baad322d-9929-0410-88f0-f92e8ff3e1bd

+++ channels/trunk/src/main/java/com/mindalliance/channels/dao/Memory.java
@@ -345,24 +345,26 @@ public final class Memory implements Dao {
      * {@inheritDoc}
      */
     public void onAfterCommand( Command command ) {
-        // Update snapshot if required
-        if ( journal.size() >= snapshotThreshold ) {
-            try {
-                takeSnapshot();
-                getJournalFile().delete();
-                journal.reset();
-            } catch ( IOException e ) {
-                throw new RuntimeException( "Failed to take snapshot", e );
+        if ( command.isMemorable() ) {
+            // Update snapshot if required
+            if ( journal.size() >= snapshotThreshold ) {
+                try {
+                    takeSnapshot();
+                    getJournalFile().delete();
+                    journal.reset();
+                } catch ( IOException e ) {
+                    throw new RuntimeException( "Failed to take snapshot", e );
 
+                }
             }
-        }
-        // else save journaled commands
-        else {
-            journal.addCommand( command );
-            try {
-                saveJournal();
-            } catch ( IOException e ) {
-                throw new RuntimeException( "Failed to save journal", e );
+            // else save journaled commands
+            else {
+                journal.addCommand( command );
+                try {
+                    saveJournal();
+                } catch ( IOException e ) {
+                    throw new RuntimeException( "Failed to save journal", e );
+                }
             }
         }
     }
